gateway/requirements.txt
Flask
Flask-CORS
requests
PyJWT
python-dotenv
